assign Schema to a particular Context. encode typeid to enable polymorphysm  #11
