+++ src/spec.ts
@@ -1,3 +1,4 @@
 export const END_OF_STRUCTURE = 0xc1; // (msgpack spec: never used)
 export const NIL = 0xc0;
 export const INDEX_CHANGE = 0xd4;
+export const TYPE_ID = 0xd5;
